Rename use-setting to use-settings
