+++ packages/block-editor/src/components/colors-gradients/use-multiple-origin-colors-and-gradients.js
@@ -7,7 +7,7 @@ import { _x } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { useSettings } from '../use-setting';
+import { useSettings } from '../use-settings';
 
 /**
  * Retrieves color and gradient related settings.
