+++ packages/block-editor/src/components/font-family/index.js
@@ -7,7 +7,7 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { useSettings } from '../use-setting';
+import { useSettings } from '../use-settings';
 
 export default function FontFamilyControl( {
 	value = '',
