+++ packages/block-editor/src/components/index.native.js
@@ -59,7 +59,7 @@ export { default as BlockCaption } from './block-caption';
 export { default as Caption } from './caption';
 export { default as PanelColorSettings } from './panel-color-settings';
 export { default as __experimentalPanelColorGradientSettings } from './colors-gradients/panel-color-gradient-settings';
-export { useSettings, default as useSetting } from './use-setting';
+export { useSettings, default as useSetting } from './use-settings';
 export {
 	RecursionProvider as __experimentalRecursionProvider,
 	useHasRecursion as __experimentalUseHasRecursion,
