+++ packages/block-editor/src/components/spacing-sizes-control/input-controls/spacing-input-control.js
@@ -20,7 +20,7 @@ import { settings } from '@wordpress/icons';
 /**
  * Internal dependencies
  */
-import { useSettings } from '../../use-setting';
+import { useSettings } from '../../use-settings';
 import { store as blockEditorStore } from '../../../store';
 import {
 	ALL_SIDES,
